app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
annotated-types==0.5.0
    # via pydantic
anyio==3.7.1
    # via
    #   fastapi
    #   starlette
    #   watchfiles
bcrypt==4.0.1
    # via -r requirements.in
black==23.9.1
    # via -r requirements.in
build==1.0.3
    # via pip-tools
click==8.1.7
    # via
    #   black
    #   pip-tools
    #   uvicorn
fastapi==0.103.1
    # via -r requirements.in
flake8==6.1.0
    # via -r requirements.in
greenlet==2.0.2
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
httptools==0.6.0
    # via uvicorn
idna==3.4
    # via anyio
mccabe==0.7.0
    # via flake8
mypy==1.5.1
    # via -r requirements.in
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
packaging==23.1
    # via
    #   black
    #   build
pathspec==0.11.2
    # via black
pip-tools==7.3.0
    # via -r requirements.in
platformdirs==3.10.0
    # via black
psycopg==3.1.10
    # via -r requirements.in
pycodestyle==2.11.0
    # via flake8
pydantic==2.3.0
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.6.3
    # via pydantic
pyflakes==3.1.0
    # via flake8
pyjwt==2.8.0
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via build
python-dotenv==1.0.0
    # via uvicorn
pyyaml==6.0.1
    # via uvicorn
sniffio==1.3.0
    # via anyio
sqlalchemy==2.0.20
    # via -r requirements.in
starlette==0.27.0
    # via fastapi
typing-extensions==4.7.1
    # via
    #   fastapi
    #   mypy
    #   psycopg
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
uvicorn[standard]==0.23.2
    # via -r requirements.in
uvloop==0.17.0
    # via uvicorn
watchfiles==0.20.0
    # via uvicorn
websockets==11.0.3
    # via uvicorn
wheel==0.41.2
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
